+++ app/javascript/components/ExpensesPreviews/index.test.js
@@ -17,76 +17,36 @@ describe("ExpensesPreviews", () => {
       travel_expenses_limit: undefined,
       food_expenses_limit: undefined,
       other_expenses_limit: undefined,
-      receipts_required: undefined,
+      receipts_required: false,
       edit_links: {}
     };
     mountedExpensesPreviews = undefined;
-  });
-
-  describe("'yes' has been selected", () => {
-    it("renders as many sections as areas on the preview that could change", () => {
-      const sections = expensesPreviews().find("section");
-      expect(sections.length).to.eql(1);
-    });
-
-    it("has blank outputs", () => {
-      expensesPreviews()
-        .find("output")
-        .forEach(output => expect(output.text()).to.be.empty);
-    });
-  });
-
-  describe("'no' has been selected", () => {
-    beforeEach(() => {
-      props = {
-        expenses_enabled: false
-      };
-    });
-    it("doesn't render any sections", () => {
-      const sections = expensesPreviews().find("section");
-      expect(sections.length).to.eql(0);
-    });
   });
 
   describe("the initial render state", () => {
     describe("props for the finalPreview are given", () => {
       beforeEach(() => {
         props = {
-          expenses_enabled: true,
-          travel_expenses_limit: "5.00",
-          food_expenses_limit: "10.00",
-          other_expenses_limit: "15.00",
+          travel_expenses_limit: "5",
+          food_expenses_limit: "10",
+          other_expenses_limit: "15",
           receipts_required: "true",
-          edit_links: {
+          finalPreview: true,
+          editLinks: {
             travel_expenses_limit: "/rails/path/to/travel_expenses_limit",
-            food_expenses_limit: "/rails/path/to/food_expenses_limit",
-            other_expenses_limit: "/rails/path/to/other_expenses_limit",
             receipts_required: "/rails/path/to/receipts_required"
           }
         };
       });
 
-      it("renders an editLink with the supplied rails route for 'Travel Expenses Limit'", () => {
-        expect(expensesPreviews()).to.contain(
-          <a className="editable" href="/rails/path/to/travel_expenses_limit">
-            5.00
-          </a>
-        );
+      it("renders an editLink with the supplied rails route for Expenses", () => {
+        expect(
+          expensesPreviews()
+            .find(".editable")
+            .first()
+        ).to.attr("href", "/rails/path/to/travel_expenses_limit");
       });
-      it("renders an editLink with the supplied rails route for 'Food Expenses Limit'", () => {
-        expect(expensesPreviews()).to.contain(
-          <a className="editable" href="/rails/path/to/food_expenses_limit">
-            15.00
-          </a>
-        );
-      });
-      it("renders an editLink with the supplied rails route for 'Other Expenses Limit'", () => {
-        expect(expensesPreviews()).to.contain(
-          <a className="editable" href="/rails/path/to/other_expenses_limit">
-            15.00
-          </a>
-        );
-      });
+
       it("renders an editLink with the supplied rails route for 'Receipts Required'", () => {
         expect(expensesPreviews()).to.contain(
           <a className="editable" href="/rails/path/to/receipts_required">
@@ -108,11 +68,32 @@ describe("ExpensesPreviews", () => {
       beforeEach(() => {
         props = {
           expenses_enabled: true,
-          travel_expenses_limit: "2.00"
+          travel_expenses_limit: "2.00",
+          receipts_required: "true"
         };
       });
 
       describe("receipt of an input change", () => {
+        it("adds an additional expense", () => {
+          expensesPreviews()
+            .instance()
+            .handleTextOrRadioChange({
+              target: {
+                name: "research_session[food_expenses_limit]",
+                value: "10.00"
+              }
+            });
+
+          expect(
+            expensesPreviews()
+              .find('output[data-field="travel_expenses_limit"]')
+              .first()
+              .text()
+          ).to.eql(
+            "We allow travel expenses of up to £2.00, and food expenses of up to £10.00."
+          );
+        });
+
         it("changes the travel expenses limit", () => {
           expensesPreviews()
             .instance()
@@ -125,10 +106,27 @@ describe("ExpensesPreviews", () => {
 
           expect(
             expensesPreviews()
-              .find('output[data-field="location"]')
+              .find('output[data-field="travel_expenses_limit"]')
               .first()
               .text()
-          ).to.eql("20.00");
+          ).to.eql("We allow travel expenses of up to £20.00.");
+        });
+
+        it("changes that receipts are required", () => {
+          expensesPreviews()
+            .instance()
+            .handleTextOrRadioChange({
+              target: {
+                name: "research_session[receipts_required]",
+                value: "false"
+              }
+            });
+
+          expect(expensesPreviews()).not.to.contain(
+            <a className="editable" href="/rails/path/to/receipts_required">
+              Receipts must be provided.
+            </a>
+          );
         });
       });
     });