+++ spec/presenters/research_session_presenter_spec.rb
@@ -102,55 +102,6 @@ RSpec.describe ResearchSessionPresenter do
     end
   end
 
-  describe '#expenses_sentence' do
-    subject { presenter.expenses_sentence }
-
-    context 'no expenses given (all nil)' do
-      let(:research_session) { build_stubbed :research_session, :nil_expenses }
-      it { is_expected.to be_nil }
-    end
-
-    context 'no expenses given (all zero)' do
-      let(:research_session) { build_stubbed :research_session, :zero_expenses }
-      it { is_expected.to be_nil }
-    end
-
-    context 'One expense is given' do
-      let(:research_session) do
-        build_stubbed :research_session, :step_where_when, travel_expenses_limit: 50.00
-      end
-      it { is_expected.to eql('We allow travel expenses of up to £50.00.') }
-    end
-
-    context 'Two expenses are given' do
-      let(:research_session) do
-        build_stubbed :research_session, :step_where_when,
-                      travel_expenses_limit: 50.00,
-                      food_expenses_limit: 10.00
-      end
-      it do
-        is_expected.to eql(
-          'We allow travel expenses of up to £50.00 and food expenses of up to £10.00.'
-        )
-      end
-    end
-
-    context 'Three expenses are given' do
-      let(:research_session) do
-        build_stubbed :research_session, :step_where_when,
-                      travel_expenses_limit: 50.00,
-                      food_expenses_limit: 10.00,
-                      other_expenses_limit: 5.00
-      end
-      it do
-        is_expected.to eql(
-          'We allow travel expenses of up to £50.00, food expenses of up to £10.00, '\
-          'and other expenses of up to £5.00.'
-        )
-      end
-    end
-  end
-
   describe '#incentive_text' do
     subject(:text) { presenter.incentive_text }